Refactor generative trainer and models
