+++ tests/test_trainer.py
@@ -2,13 +2,8 @@ import torch
 from torch.utils.data import DataLoader
 from xtylearner.data import load_toy_dataset, load_mixed_synthetic_dataset
 from xtylearner.models import CycleDual, MixtureOfFlows, MultiTask
-from xtylearner.training import (
-    SupervisedTrainer,
-    M2VAE,
-    SS_CEVAE,
-    M2VAETrainer,
-    CEVAETrainer,
-)
+from xtylearner.training import SupervisedTrainer, GenerativeTrainer
+from xtylearner.models import M2VAE, SS_CEVAE
 
 
 def test_supervised_trainer_runs():
@@ -49,7 +44,7 @@ def test_m2vae_trainer_runs():
     loader = DataLoader(dataset, batch_size=5)
     model = M2VAE(d_x=2, d_y=1, k=2)
     opt = torch.optim.Adam(model.parameters(), lr=0.01)
-    trainer = M2VAETrainer(model, opt, loader)
+    trainer = GenerativeTrainer(model, opt, loader)
     trainer.fit(1)
     loss = trainer.evaluate(loader)
     assert isinstance(loss, float)
@@ -60,7 +55,7 @@ def test_cevae_trainer_runs():
     loader = DataLoader(dataset, batch_size=5)
     model = SS_CEVAE(d_x=2, d_y=1, k=2)
     opt = torch.optim.Adam(model.parameters(), lr=0.01)
-    trainer = CEVAETrainer(model, opt, loader)
+    trainer = GenerativeTrainer(model, opt, loader)
     trainer.fit(1)
     loss = trainer.evaluate(loader)
     assert isinstance(loss, float)
@@ -82,7 +77,7 @@ def test_cevae_trainer_mixed_dataset():
     loader = DataLoader(dataset, batch_size=5)
     model = SS_CEVAE(d_x=2, d_y=1, k=2)
     opt = torch.optim.Adam(model.parameters(), lr=0.01)
-    trainer = CEVAETrainer(model, opt, loader)
+    trainer = GenerativeTrainer(model, opt, loader)
     trainer.fit(1)
     loss = trainer.evaluate(loader)
     assert isinstance(loss, float)
