+++ xtylearner/__init__.py
@@ -1,7 +1,6 @@
 """Core package for XTYLearner models and training utilities."""
 
-from .models import CycleDual, MixtureOfFlows, MultiTask, get_model
-from .training import M2VAE, SS_CEVAE
+from .models import CycleDual, MixtureOfFlows, MultiTask, get_model, M2VAE, SS_CEVAE
 
 __all__ = [
     "CycleDual",
