+++ xtylearner/models/__init__.py
@@ -3,11 +3,14 @@
 from .cycle_dual import CycleDual
 from .flow_ssc import MixtureOfFlows
 from .multitask_selftrain import MultiTask
+from .generative import M2VAE, SS_CEVAE
 from .registry import get_model
 
 __all__ = [
     "CycleDual",
     "MixtureOfFlows",
     "MultiTask",
+    "M2VAE",
+    "SS_CEVAE",
     "get_model",
 ]
